refactoring
- add file with `ConstSpeciesAttributes.hpp` for charge and mass
- add file `openPMDDimension.hpp` with conversion helper `PIConGPU dimension`
to `openPMD-api dimension`
